Adding an interface to fail individual wells
It's possible to fail individual wells by selecting them and updating
the plate.

Did a couple of tidy ups so that plates that need no interaction don't
display the 'Create Plate' button and just jump through to being
created.

Added a view of the flash messages so that people know things have
happened.