Remove qc complete step from general plate workflow
This step was of little use, and actually restricted
a few common requirements. IN addition it was increasing code
complexity.

I've left some of the basic support in for existing plates
in the state, as well as a little of the advanced behaviour
for handling it. This should be removed at a later date.