Pre-capture pools are sorted in column order
Pre-capture pools in the ISC pipeline were no-longer being sorted in
column order. This was causing delays in the lab.

While hashes in ruby are ordered based on insert order, there is no
such provision in javascript objects. Thus the browser does not have
to maintain key order when loading a JSON object. Indeed, it appears
that both chomium and firefox are sorting keys in alphabetical order.

This bug has been longstanding, but has mostly been hidden as the
pool-ids were typically generated in column order. In the plates that
were causing the problems, this was not the case. This could be due to
changes in SSR behaviour, or more likely the asset refactor has changed
the order in which a query is sorted during submission processing.

This fix:
- Updates the helper function to return an array, not a hash to ensure
  order is maintained.
- Ensures the javascript maintains the data in an array format
- Changes the format of the data output by the helper function to reduce
  the need for subsequent processing by the javascript
- Renames some of the variable to better reflect what they actually
  contain
- Removes code associated with legacy behaviour which is no longer
  required
    * No need to pass through pool ids
    * No need to handle nested submission/pre-cap pool information
- Adapt coding style to meet current linter expectations