Halve time taken to run rspec suite
I hoped the update of the json_client_api gem would allow
replacing my mocks with simple asignments. Unfortunately things don't
appear to be that simple, even though a few issues imply they should be
now.

In investigating why things didn't work however, I discovered an
alternative approach, which is substantially faster.

A handful of Mocks still remain in a few of the more complicated corners
but I need to move on now, so these remain for now.