Rubocop driven refactor
We a lot of shared requests behaviour
between plates and wells. This make plates conform
to the interface required for 'HasRequests'.
Plate still need a custom active_requests
method to maintain the previous behaviour
but this is probably something we could change in future.

Rather than a module this could be adapted
to a proxy class. I had a bit of a play around to
do this, but the differing active_requests behaviour
made it not entirely straight forward, so I don't want to rush it.