Use yaml anchors for shared filters
Fixes ##557

Each pipeline has a 'filters' parameter which determines which requests it is active for. For branching pipeline, the same filters parameter may appear on multiple branches of the pipeline. This redundancy has resulted in a few bugs where a new library_type has only been added to one half of the pipeline.

This uses yaml anchors to remove redundancy.

I've validated that the parse yaml files are identical.