Stock plate factory passed by default
Almost all uses of the factory were
setting the state to 'passed' and this
represents the state usually seen by limber.
In contrast 'pending' only happens if
no submission has been made.

I've also switched a couple of stock plates
to standard plates, as it better reflects
the expected role.