Apply sort order to stock plate retrieval
Stock plate ordering was not specified, so was dependant on the
behaviour of the sequencescape database. This was resulting in
stock plates appearing to change.

This brings the behaviour mostly in line with Sequencescape, although
uses id over created at. This has the advantage of higher resolution,
quicker speed, and not needing a new sort scope added.

That said, we should probably make the same change within sequencescape.

Fixes #619