Disable cross plate pool detection for LTHR-384
Fixes #647

The consolidation at the beginning of the LTHR pipeline
means we aren't actually going to be pooling with other plates
in the pool. This just means that we'll unnecessarily block rework
loops.

While a smarter solution would be great, it hides a surprising amount
of complexity and edge cases (see the linked issue) so I've opted
for an explicit disable.