Fix state changer using wrong reasons
Previously the state changer was trying to use javascript
to disable the reason fields for the opposing state change.
However this had actually been implemented incorrectly, and also
ignored the customer accepts responsibility checkbox.

Generally I felt this approach was brittle, and could result in
unexpected behaviour if the javascript failed, or a developer added new
fields.

Instead we group each set of fields with 'fields_for' and select them
in the controller.

This greatly simplifies the javascript, and lets us remove some helper
methods. Their is a slight cost increase in the controller.