Fix font scaling on tags
Font size smaller is abolute, so was making the tags
too large on 384 well plates.

The change was made to handle the chromium pipeline,
which has tag indexes of 3 digits long. These have trouble
fitting on the well.

Meanwhile, while a scale of 80% works, it makes the text
uncomfortably small on 384 well plates.

Instead we only scale the text down when necessary,
keeping everything nice and readable.

I did look for a way to get the css to handle this
automatically, but didn't have much luck.