Rubocop - Class length
We ride so close to the limit here, that almost any
change requires a pay off elesewhere. I'm a bit torn on where we go from here:

1) Pulling functionality onto separate objects,
which is definitely the intent of the cop.
2) Add a disable here.

I think there is still some reasonable scope for
1, but we need to make sure were making actual
improvements, not just appeasing the cops.