Allow rake task to fetch rack purposes
- Fetch all purposes, not just tube and plate
- Migrate to API V2 to fetch this

We get a major speed improvement here, as
purposes no longer need to count(*) all their
plates.

- Adjust pagination and combination

The base paginator wasn't correctly handling
the pages parameters, but the strict nested
params paginator works a treat. This allows
us to tidy up the fetching of all pages to be a
little more idiomatic