Refactor purpose config ahead of tube racks
The default hashes were being mutated during loading
which could lead to hard to diagnose issues later. I've refactored,
taking care to ensure that the resultant config yamls are equal.

Note: There *will* be a change of sha, as key order will differ.
However we are not dependent on hash ordering here.

There's definitely scope for further refactoring here, as the printer
stuff is a little confusing, and in places is badly named.