Quick and dirty plate scaling
Hard coding the plate scaling to fit. However
this isn't suaitable long term, so we'll need
to dynamically scale the plate.

Unfortunately I couldn't find any sloutions that
wouldn't depend on javascript. While it is possible
to get cell content to wrap, this pushes the
plate height too high and compromises readability.