Use media queries to select style
We're fortunate that trhe two print styles are dependent on
whether we are printing portrait
or landscape, so we can use a simple media query.

This is superior to the attempt
to use alternative stylesheets because:
1) It requires no custom js, and
can make use of the native print button.
2) Altertnative style-sheet support wasn't great, and not only
did we need to build half the support
ourselves, but we were building on
a shaky foundation which might be
removed any moment.