Switch to zeitwerk in preparation for Rails 7
Note: The change in naming of GenerateQCResults could also have been
accomplished by adding QC as an acronym. eg.

```ruby
ActiveSupport::Inflector.inflections(:en) do |inflect|
  inflect.acronym "VAT"
end
```

However this impact the naming of more files, including
Limber::QcResults, which is a subclass of the Sequencescape::QcResults
imported from the V1 API. I'm not sure if renaming this would allow
the correct class to be initialized, so keeping things low risk at the
moment.