Reinitialize stettings on reload
Previously we would have failed to
reload the PipelineList, using the
original version. While this did work,
it could be a bit confusing in future.

Plus this approach lets us hot reload
the stettings, although we will still
need to run rake config:generate. However
the performance gains from this make the
lack of dynamic reloading acceptable.

In future we could consider hashing the various config files, and rerunning
setup in the event of changes.